refactor(app): move binary tests out of main
